Configure Watchman during ConfigStep
